+++ Application/Sources/Application Services/KeyboardManagerApplicationService.swift
@@ -12,7 +12,12 @@ import IQKeyboardManagerSwift
 final class KeyboardManagerApplicationService: NSObject, ApplicationService {
 	// swiftlint:disable:next discouraged_optional_collection
 	func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-		IQKeyboardManager.shared.enable = true
+		IQKeyboardManager.shared.enable = false
+
+		// only enable for specific screens
+		IQKeyboardManager.shared.enabledDistanceHandlingClasses.append(contentsOf: [
+			ViewController.self
+		])
 
 		return true
 	}