+++ Application/Sources/Helpers/AppDelegate.swift
@@ -12,8 +12,12 @@ import AppwiseCore
 final class AppDelegate: AppwiseCore.AppDelegate<Config> {
 	override var services: [ApplicationService] {
 		return [
+			CoreDataApplicationService(),
 			FabricApplicationService(),
-			KeyboardManagerApplicationService()
+			KeyboardManagerApplicationService(),
+			NetworkDebugApplicationService(),
+			StyleApplicationService(),
+			PushNotificationsApplicationService()
 		]
 	}
 }